+++ app/src/main/res/layout/fragment_day_pager.xml
