+++ app/src/main/res/menu/bottom_navigation_menu.xml
@@ -1,0 +1,11 @@
+<?xml version="1.0" encoding="utf-8"?>
+<menu xmlns:android="http://schemas.android.com/apk/res/android">
+    <item
+        android:id="@+id/scoreboardFragment"
+        android:icon="@drawable/ic_basketball"
+        android:title="@string/games" />
+    <item
+        android:id="@+id/fantasyFragment"
+        android:icon="@drawable/ic_table"
+        android:title="@string/fantasy" />
+</menu>
